move some of the photo page to a generic component
