+++ package.json
@@ -18,7 +18,6 @@
     "dependencies": {
         "@headlessui/react": "^1.7.14",
         "@heroicons/react": "^2.0.17",
-        "@next/font": "^13.2.4",
         "classnames": "^2.3.2",
         "cloudinary": "^1.36.4",
         "date-fns": "^2.29.3",
